# openapi/teaching.yaml
openapi: 3.1.0
info:
  title: Sales Management — Module giảng dạy (Nhân sự & Giáo viên)
  version: 1.0.0
  description: |
    Quản lý giáo viên, lịch dạy (mẫu lặp theo tuần) và chấm công theo từng buổi.

    Role: `nhansu` toàn quyền; `director`/`director_la`/`troly_gd`/`ketoan_truong` chỉ xem;
    `giaovien` chỉ xem dữ liệu của chính mình. Phạm vi luôn suy ra từ access token.

    Dự án chưa cài `@nestjs/swagger` nên spec này được duy trì thủ công.
    Chi tiết: TEACHING-MODULE-API.md
servers:
  - url: http://localhost:3010
    description: production (pm2 sales-be)
  - url: http://localhost:3011
    description: development
...
security:
  - bearerAuth: []
paths:
  /school-classes:
    get:
      tags: [school-classes]
      operationId: listSchoolClasses
      summary: Danh sách lớp học
      description: Role `giaovien` đọc được để hiển thị tên lớp, nhưng không ghi.
      parameters:
        - { name: schoolId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: schoolYear, in: query, schema: { type: string, maxLength: 20 } }
        - { name: gradeLevel, in: query, schema: { type: integer, minimum: 1, maximum: 12 } }
        - name: catalogId
          in: query
          description: >-
            Môn trong danh mục dùng chung. Có tham số này thì mỗi lớp được trả kèm
            subjectStatus / subjectId / subjectName / subjectReason — môn tương ứng
            của trường, để xem trước lớp nào áp được môn.
          schema: { type: integer, minimum: 1 }
        - { name: isActive, in: query, schema: { type: boolean } }
        - name: search
          in: query
          description: Theo tên lớp hoặc GVCN, không phân biệt hoa/thường
          schema: { type: string, maxLength: 100 }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: OK — sắp xếp sẵn theo khối rồi tên lớp
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/SchoolClass' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [school-classes]
      operationId: createSchoolClass
      summary: Tạo lớp học cho một trường (nhansu)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSchoolClass' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchoolClass' }
        '400':
          description: Thiếu tên lớp/năm học, hoặc trường không tồn tại
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Trùng tên lớp trong cùng trường + năm học
  /school-classes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      tags: [school-classes]
      operationId: getSchoolClass
      summary: Chi tiết lớp học
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchoolClass' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Lớp học không tồn tại }
    patch:
      tags: [school-classes]
      operationId: updateSchoolClass
      summary: Sửa lớp học (nhansu)
      description: '`schoolId` bị bỏ qua — không đổi được trường của lớp.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSchoolClass' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchoolClass' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Lớp học không tồn tại }
        '409': { description: Trùng tên lớp trong cùng trường + năm học }
    delete:
      tags: [school-classes]
      operationId: deleteSchoolClass
      summary: Xoá lớp học (nhansu)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
                  id: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Lớp học không tồn tại }
        '409':
          description: Lớp đã có lịch/buổi dạy — hãy đặt isActive = false
...
  /teaching-schedules:
    get:
      tags: [teaching-schedules]
      operationId: listTeachingSchedules
      summary: Danh sách mẫu lịch lặp
      parameters:
        - { name: teacherId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: schoolId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: classId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: subjectId, in: query, schema: { type: integer, minimum: 1 } }
        - name: dayOfWeek
          in: query
          description: 2 = Thứ Hai … 7 = Thứ Bảy, 8 = Chủ Nhật
          schema: { type: integer, minimum: 2, maximum: 8 }
        - { name: isActive, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TeachingSchedule' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
    post:
      tags: [teaching-schedules]
      operationId: createTeachingSchedule
      summary: Tạo mẫu lịch lặp theo tuần (role nhansu)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTeachingSchedule' }
            example:
              teacherId: 1
              schoolId: 494
              subjectId: 794
              dayOfWeek: 3
              startTime: '07:30'
              endTime: '09:00'
              effectiveFrom: '2026-08-01'
              effectiveTo: '2026-08-31'
      responses:
        '201': { description: Đã tạo }
        '400':
          description: Giờ/ngày không hợp lệ, môn không thuộc trường, giáo viên ngừng hoạt động
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Giáo viên đã có lịch trùng thứ + giao giờ + giao khoảng hiệu lực
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                message: Giáo viên đã có lịch Thứ Ba 07:30–09:00 tại TIỂU HỌC THẮNG NHÌ trong khoảng thời gian này
                error: Conflict
                statusCode: 409
...
  /teaching-schedules/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [teaching-schedules]
      operationId: getTeachingSchedule
      summary: Chi tiết mẫu lịch
      responses:
        '200': { description: OK }
        '404': { description: Không tồn tại }
    patch:
      tags: [teaching-schedules]
      operationId: updateTeachingSchedule
      summary: Sửa mẫu lịch (role nhansu)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTeachingSchedule' }
      responses:
        '200': { description: OK }
        '409': { description: Trùng lịch giáo viên }
    delete:
      tags: [teaching-schedules]
      operationId: deleteTeachingSchedule
      summary: Xoá mẫu lịch (role nhansu)
      description: Buổi đã sinh bị xoá theo. 409 nếu đã có buổi được chấm công.
      responses:
        '200': { description: 'Đã xoá: { "deleted": true }' }
        '409': { description: Đã có buổi được chấm công }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Thiếu hoặc sai token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Role không có quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: Forbidden resource, error: Forbidden, statusCode: 403 }
  schemas:
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
    SchoolClass:
      type: object
      properties:
        id: { type: integer }
        schoolId: { type: integer }
        schoolName: { type: string }
        name: { type: string, example: 1A }
        gradeLevel: { type: [integer, 'null'], description: Khối 1–12; null với mầm non }
        schoolYear: { type: string, example: 2026-2027 }
        studentCount: { type: integer }
        homeroomTeacher:
          type: [string, 'null']
          description: GVCN phía trường — chỉ là tên, không phải hồ sơ `teachers`
        isActive: { type: boolean }
        note: { type: [string, 'null'] }
        scheduleCount: { type: integer, description: Số mẫu lịch đang dùng lớp này }
        sessionCount: { type: integer, description: Số buổi dạy đang dùng lớp này }
        subjectStatus:
          type: string
          enum: [RESOLVED, MISSING, AMBIGUOUS]
          description: Chỉ có khi truy vấn kèm catalogId
        subjectId: { type: [integer, 'null'], description: Môn của trường khi RESOLVED }
        subjectName: { type: [string, 'null'] }
        subjectReason: { type: [string, 'null'], description: Lý do khi không RESOLVED }
    CreateSchoolClass:
      type: object
      required: [schoolId, name, schoolYear]
      properties:
        schoolId: { type: integer, minimum: 1 }
        name:
          type: string
          maxLength: 100
          description: Tự bỏ khoảng trắng thừa; không trùng trong cùng trường + năm học
        schoolYear:
          type: string
          maxLength: 20
          description: Chuỗi tự do, cùng quy ước với subjects.school_year
          example: 2026-2027
        gradeLevel: { type: integer, minimum: 1, maximum: 12 }
        studentCount: { type: integer, minimum: 0, maximum: 1000, default: 0 }
        homeroomTeacher: { type: string, maxLength: 255 }
        isActive: { type: boolean, default: true }
        note: { type: string, maxLength: 1000 }
    UpdateSchoolClass:
      type: object
      description: Mọi field tuỳ chọn; `schoolId` gửi lên sẽ bị bỏ qua.
      properties:
        name: { type: string, maxLength: 100 }
        schoolYear: { type: string, maxLength: 20 }
        gradeLevel: { type: [integer, 'null'], minimum: 1, maximum: 12 }
        studentCount: { type: integer, minimum: 0, maximum: 1000 }
        homeroomTeacher: { type: string, maxLength: 255 }
        isActive: { type: boolean }
        note: { type: string, maxLength: 1000 }
    CreateTeachingSchedule:
      type: object
      required: [teacherId, classId, subjectId, dayOfWeek, startTime, endTime, effectiveFrom]
      properties:
        teacherId: { type: integer, minimum: 1 }
        classId:
          type: integer
          minimum: 1
          description: Lịch được xếp cho lớp; trường suy ra từ lớp
        schoolId:
          type: integer
          minimum: 1
          description: Tuỳ chọn — chỉ để đối chiếu, phải đúng trường của classId
        subjectId:
          type: integer
          minimum: 1
          description: Phải thuộc trường của classId, nếu không trả 400
        dayOfWeek:
          type: integer
          minimum: 2
          maximum: 8
          description: 2 = Thứ Hai … 7 = Thứ Bảy, 8 = Chủ Nhật
        startTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d(:[0-5]\d)?$' }
        endTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d(:[0-5]\d)?$' }
        effectiveFrom: { type: string, format: date }
        effectiveTo:
          type: string
          format: date
          description: Bỏ trống = chưa có ngày kết thúc
        ratePerPeriod:
          type: [number, 'null']
          minimum: 0
          maximum: 100000000
          description: >-
            Đơn giá mỗi tiết (VND), tối đa 2 chữ số thập phân. Bỏ trống = kế thừa
            (mẫu lịch -> giáo viên). 0 là giá hợp lệ và KHÔNG kế thừa.
        isActive: { type: boolean, default: true }
        note: { type: string, maxLength: 1000 }
    TeachingSchedule:
      type: object
      properties:
        id: { type: integer }
        teacherId: { type: integer }
        teacherName: { type: string }
        schoolId: { type: integer }
        schoolName: { type: string }
        classId: { type: [integer, 'null'], description: null = lịch cũ chưa gắn lớp }
        className: { type: [string, 'null'], example: 1A }
        classGradeLevel: { type: [integer, 'null'] }
        subjectId: { type: integer }
        subjectName: { type: string }
        schoolYear: { type: [string, 'null'] }
        dayOfWeek: { type: integer }
        dayOfWeekLabel: { type: [string, 'null'], example: Thứ Ba }
        startTime: { type: string, example: '07:30' }
        endTime: { type: string, example: '09:00' }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        ratePerPeriod:
          type: [number, 'null']
          description: Đơn giá mỗi tiết đã chốt cho mẫu này; null = chưa khai
        isActive: { type: boolean }
        note: { type: [string, 'null'] }
...
    Error:
      type: object
      properties:
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string }
        statusCode: { type: integer }